Display pb-answer blocks in edit view for staff user to select.
We are using the edX Course Blocks API to fetch all pb-answer blocks in
the current course
All pb-answer blocks in the course are displayed as labeled checkboxes
in the studio edit view.
The checkboxes are arranged by section/subsection/unit in the order as
they appear in the course.