[BB-2132] [YONK-1560] Support JWT creation on openedx ironwood (#27)
* Support JWT creation on openedx ironwood

Problem root cause: JwtBuilder class was removed in commit dc56a63e of edx-platform
git repository and the eoc-journal xblock was not updated accordingly.

This patch drops support of all openedx versions before ironwood.

* Set pdfminer version in Python requirements for test environment

pdfminer 20140328 is the last version to support Python 2, used by
EDX platform. Due to this, pdfminer newer versions (20091010 and later)
were breaking the tests.

* Fix tests broken by openedx import

An import of Python package openedx was added in commit 4507e0d, but
the test environment does not have that Python package. To avoid
breaking the tests, this patch:

- ignores the import error in Pylint check.
- move the import to inside the function, which is not called in tests
  because the methods which call it are already mocked

* Create a Python module which contains edx-platform dependent code

* Version bump